experiments/fl-iot-ids-v2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fl-iot-ids-v2"
version = "0.2.0"
description = "Robust Federated IDS benchmark for IoT using Flower"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Saif Eddinne Boukhatem" }
]
dependencies = [
  "flwr[simulation]>=1.22.0",
  "torch>=2.2.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.5.0",
  "protobuf>=5.29,<6",
  "pyyaml>=6.0.1",
  "pydantic>=2.7.0",
  "mlflow>=2.12.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "prometheus-client>=0.20.0",
  "rich>=13.7.0"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v"
